rattler_pty::unix

Struct PtyProcess

pub struct PtyProcess {
    pub pty: PtyMaster,
    pub child_pid: Pid,
    /* private fields */
}

Start a process in a forked tty so you can interact with it the same as you would within a terminal

The process and pty session are killed upon dropping PtyProcess

Fields
pty: PtyMasterchild_pid: Pid
Implementations

impl PtyProcess

pub fn new(command: Command, opts: PtyProcessOptions) -> Result<Self>

Start a process in a forked pty

pub fn get_file_handle(&self) -> Result<File>

Get handle to pty fork for reading/writing

pub fn status(&self) -> Option<WaitStatus>

Get status of child process, non-blocking.

This method runs waitpid on the process. This means: If you ran exit() before or status() this method will return None

pub fn exit(&mut self) -> Result<WaitStatus>

Regularly exit the process, this method is blocking until the process is dead

pub fn kill(&mut self, sig: Signal) -> Result<WaitStatus>

Kill the process with a specific signal. This method blocks, until the process is dead

repeatedly sends SIGTERM to the process until it died, the pty session is closed upon dropping PtyMaster, so we don’t need to explicitly do that here.

if kill_timeout is set and a repeated sending of signal does not result in the process being killed, then kill -9 is sent after the kill_timeout duration has elapsed.

pub fn set_raw(&self) -> Result<Termios>

Set raw mode on stdin and return the original mode

pub fn set_mode(&self, original_mode: Termios) -> Result<()>

pub fn set_window_size(&self, window_size: Winsize) -> Result<()>

pub fn kill_timeout(&self) -> Option<Duration>

pub fn set_kill_timeout(&mut self, timeout: Option<Duration>)

When calling kill() or async_kill(), if the process doesn’t respond to the signal within this timeout, it will be forcefully killed with SIGKILL.

pub async fn async_read(&self, buf: &mut [u8]) -> Result<usize>

pub async fn async_write(&self, buf: &[u8]) -> Result<usize>

pub async fn async_wait(&mut self) -> Result<WaitStatus>

pub async fn async_exit(&mut self) -> Result<WaitStatus>

pub async fn async_kill(&mut self, sig: Signal) -> Result<WaitStatus>