rattler_pty::unix

Struct PtySession

pub struct PtySession {
    pub process: PtyProcess,
    pub process_stdout: File,
    pub process_stdin: File,
    pub rolling_buffer: Vec<u8>,
}

Fields
process: PtyProcessprocess_stdout: File

A file handle of the stdout of the pty process

process_stdin: File

A file handle of the stdin of the pty process

rolling_buffer: Vec<u8>

Rolling buffer used for the wait_until pattern matching

Implementations

impl PtySession

use std::process::Command;
use rattler_pty::unix::PtySession;

let process = PtySession::new(Command::new("bash")).unwrap();

pub fn new(command: Command) -> Result<Self>

Constructs a new session

pub fn send<B: AsRef<[u8]>>(&mut self, s: B) -> Result<usize>

Send string to process. As stdin of the process is most likely buffered, you’d need to call flush() after send() to make the process actually see your input.

Returns number of written bytes

pub fn send_line(&mut self, line: &str) -> Result<usize>

Sends string and a newline to process. This is guaranteed to be flushed to the process. Returns number of written bytes.

pub fn flush(&mut self) -> Result<()>

Make sure all bytes written via send() are sent to the process

pub fn exit(&mut self) -> Result<WaitStatus>

Exit the process gracefully by sending SIGTERM.

This method blocks until the process has exited.

pub fn interact(&mut self, wait_until: Option<&str>) -> Result<Option<i32>>

Interact with the process. This will put the current process into raw mode and forward all input from stdin to the process and all output from the process to stdout. This will block until the process exits.